requirements.txt
beautifulsoup4==4.9.3
bidict==0.21.2
boto3==1.18.16
botocore==1.21.16
bs4==0.0.1
cachelib==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dataclasses==0.8
dnspython==1.16.0
eventlet==0.30.2
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Session==0.4.0
Flask-SocketIO==4.3.1
greenlet==1.1.1
gunicorn==20.0.2
idna==3.2
importlib-metadata==4.6.3
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.1.1
Pillow==8.3.1
pkg_resources==0.0.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
